paris::formatter

Function format_string

pub fn format_string<S>(input: S, with_colors: bool) -> String
where
    S: Into<String>,

Finds all keys in the given input. If with_colors is true, it will replace all keys with their respective ANSI color code. Otherwise it will only replace the keys with an empty string.

This function does not take into account custom styles, you need the struct for that.